back/requirements.txt
requests==2.30.0
pandas==2.1.4
xlrd==2.0.1
matplotlib==3.7.1
chardet==5.1.0
openpyxl==3.1.2
pyyaml==6.0
unicodecsv==0.14.1
Unidecode==1.3.7
psycopg2==2.9.1
python-dotenv==1.0.0
sqlalchemy==2.0.24
tqdm==4.66.1